+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changes to Joyride
 
 ## [Unreleased]
 
-- Update SCI and sci.configs to latest commits
+- Update SCI and sci.configs to latest commits; `f40ee1a6cf6c53b8271b41c101c3391773d67e76` and `aa84a1b4f1fe45735e5b748769309fc842f737c1` respectively.
+- Fix: [Output lacks trailing newlines for `println`](https://github.com/BetterThanTomorrow/joyride/pull/253)
 
 ## [0.0.68] - 2025-10-25